pyproject.toml
[tool.poetry]
name = "tlds"
version = "0.1.0"
description = ""
authors = ["Angad Behl <angad@angad.me>"]
readme = "README.md"
packages = [{ include = "tlds" }]
scripts = { tlds = "tlds.__main__:main" }

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
slack-bolt = "^1.20.0"
langchain = "^0.2.14"
langchain-community = "^0.2.12"
langchain-ollama = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"